Week 3 - Implement a Simple Data Enrichment Process/service/requirements.txt
kafka-python
pymysql
pymemcache
elasticsearch
cryptography
